Add files via upload
The project provides most of the basic functionality required for an event. It allows the user to select from
a list of event types. Once the user enters an event type eg(Marriage, Dance Show ), the system then
allows the user to select the date and time of event, place and the event equipment’s. All this data is
logged in the database and the user is given a receipt number for his booking.